+++ sdk/cosmos/azure-spring-data-cosmos-test/src/test/java/com/azure/spring/data/cosmos/repository/integration/ReactiveIndexPolicyUpdateIT.java
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 package com.azure.spring.data.cosmos.repository.integration;
 
+import com.azure.cosmos.implementation.ImplementationBridgeHelpers;
 import com.azure.cosmos.models.CosmosContainerProperties;
 import com.azure.cosmos.models.ExcludedPath;
 import com.azure.cosmos.models.IncludedPath;
@@ -12,6 +13,7 @@ import com.azure.spring.data.cosmos.core.ReactiveCosmosTemplate;
 import com.azure.spring.data.cosmos.domain.Address;
 import com.azure.spring.data.cosmos.domain.ComplexIndexPolicyEntity;
 import com.azure.spring.data.cosmos.domain.IndexPolicyEntity;
+import com.azure.spring.data.cosmos.domain.IndexPolicyOverwriteEntity;
 import com.azure.spring.data.cosmos.repository.TestRepositoryConfig;
 import com.azure.spring.data.cosmos.repository.support.CosmosEntityInformation;
 import com.azure.spring.data.cosmos.repository.support.SimpleReactiveCosmosRepository;
@@ -46,15 +48,17 @@ public class ReactiveIndexPolicyUpdateIT {
 
     CosmosEntityInformation<ComplexIndexPolicyEntity, String> complexIndexPolicyEntityInformation = new CosmosEntityInformation<>(ComplexIndexPolicyEntity.class);
 
+    CosmosEntityInformation<IndexPolicyOverwriteEntity, String> indexPolicyOverwriteEntityInformation = new CosmosEntityInformation<>(IndexPolicyOverwriteEntity.class);
+
     CosmosEntityInformation<Address, String> addressEntityInformation = new CosmosEntityInformation<>(Address.class);
 
     @Before
     public void setup() {
-        collectionManager.ensureContainersCreatedAndEmpty(template, IndexPolicyEntity.class, ComplexIndexPolicyEntity.class);
+        collectionManager.ensureContainersCreatedAndEmpty(template, IndexPolicyEntity.class, ComplexIndexPolicyEntity.class, IndexPolicyOverwriteEntity.class);
     }
 
     @Test
-    public void testIndexPolicyUpdatesOnRepoInitialization() {
+    public void testIndexPolicyDoesntUpdateOnRepoInitialization() {
         // set index policy based on entity annotation
         new SimpleReactiveCosmosRepository<>(defaultIndexPolicyEntityInformation, template);
 
@@ -84,12 +88,48 @@ public class ReactiveIndexPolicyUpdateIT {
 
         // assert
         assertThat(properties.getIndexingPolicy().getIncludedPaths().size()).isEqualTo(1);
-        assertThat(properties.getIndexingPolicy().getIncludedPaths().get(0).getPath()).isEqualTo("/field/?");
-        assertThat(properties.getIndexingPolicy().getExcludedPaths().size()).isEqualTo(2);
-        assertThat(properties.getIndexingPolicy().getExcludedPaths().get(0).getPath()).isEqualTo("/*");
-        assertThat(properties.getIndexingPolicy().getExcludedPaths().get(1).getPath()).isEqualTo("/\"_etag\"/?");
+        assertThat(properties.getIndexingPolicy().getIncludedPaths().get(0).getPath()).isEqualTo("/*");
+        assertThat(properties.getIndexingPolicy().getExcludedPaths().size()).isEqualTo(1);
+        assertThat(properties.getIndexingPolicy().getExcludedPaths().get(0).getPath()).isEqualTo("/\"_etag\"/?");
         assertThat(properties.getIndexingPolicy().isAutomatic()).isEqualTo(true);
         assertThat(properties.getIndexingPolicy().getIndexingMode()).isEqualTo(IndexingMode.CONSISTENT);
+    }
+
+    @Test
+    public void testIndexPolicyUpdatesOnRepoInitialization() {
+        // set index policy from entity annotation
+        new SimpleReactiveCosmosRepository<>(indexPolicyOverwriteEntityInformation, template);
+
+        // get original index policy
+        CosmosContainerProperties properties = template.getContainerProperties(indexPolicyOverwriteEntityInformation.getContainerName()).block();
+
+        // assert
+        assertThat(properties.getIndexingPolicy().getIncludedPaths().size()).isEqualTo(1);
+        assertThat(properties.getIndexingPolicy().getIncludedPaths().get(0).getPath()).isEqualTo("/\"_etag\"/?");
+        assertThat(properties.getIndexingPolicy().getExcludedPaths().size()).isEqualTo(1);
+        assertThat(properties.getIndexingPolicy().getExcludedPaths().get(0).getPath()).isEqualTo("/*");
+
+        // set new index policy
+        IndexingPolicy newIndexPolicy = new IndexingPolicy();
+        ImplementationBridgeHelpers.IndexingPolicyHelper.IndexingPolicyAccessor accessor =
+            ImplementationBridgeHelpers.IndexingPolicyHelper.getIndexingPolicyAccessor();
+        accessor.setOverwritePolicy(newIndexPolicy, true);
+        newIndexPolicy.setIncludedPaths(Collections.singletonList(new IncludedPath("/*")));
+        newIndexPolicy.setExcludedPaths(Collections.singletonList(new ExcludedPath("/\"_etag\"/?")));
+
+        // apply new index policy
+        CosmosEntityInformation<IndexPolicyOverwriteEntity, String> spyEntityInformation = Mockito.spy(indexPolicyOverwriteEntityInformation);
+        Mockito.doReturn(newIndexPolicy).when(spyEntityInformation).getIndexingPolicy();
+        new SimpleReactiveCosmosRepository<>(spyEntityInformation, template);
+
+        // retrieve updated index policy
+        properties = template.getContainerProperties(indexPolicyOverwriteEntityInformation.getContainerName()).block();
+
+        // assert
+        assertThat(properties.getIndexingPolicy().getIncludedPaths().size()).isEqualTo(1);
+        assertThat(properties.getIndexingPolicy().getIncludedPaths().get(0).getPath()).isEqualTo("/*");
+        assertThat(properties.getIndexingPolicy().getExcludedPaths().size()).isEqualTo(1);
+        assertThat(properties.getIndexingPolicy().getExcludedPaths().get(0).getPath()).isEqualTo("/\"_etag\"/?");
     }
 
     @Test