+++ sdk/cosmos/azure-spring-data-cosmos/src/main/java/com/azure/spring/data/cosmos/Constants.java
@@ -17,7 +17,7 @@ public final class Constants {
     /**
      * Default indexing policy automatic.
      */
-    public static final boolean DEFAULT_INDEXING_POLICY_OVERWRITE_POLICY = true;
+    public static final boolean DEFAULT_INDEXING_POLICY_OVERWRITE_POLICY = false;
 
     /**
      * Default indexing policy automatic.